+++ provision/ubuntu.sh
@@ -19,8 +19,6 @@ export JAIL_START_EXTRA="allow.mount
 export JAIL_CONF_EXTRA='
 		allow.raw_sockets;'
 export JAIL_FSTAB="
-linprocfs $ZFS_JAIL_MNT/ubuntu/compat/linux/proc    linprocfs rw,late  0 0
-linsysfs  $ZFS_JAIL_MNT/ubuntu/compat/linux/sys     linsysfs  rw,late  0 0
 #/tmp      $ZFS_JAIL_MNT/ubuntu/compat/linux/tmp     nullfs    rw,late  0 0
 #/home     $ZFS_JAIL_MNT/ubuntu/compat/linux/home    nullfs    rw,late  0 0"
 
@@ -37,7 +35,6 @@ for _fs in dev proc sys tmp home; do
 done
 chmod 777 "$STAGE_MNT/compat/linux/tmp"
 assure_devfs_linux_ruleset
-configure_linux_devfs
 start_staged_jail ubuntu
 install_ubuntu
 promote_staged_jail ubuntu